requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.6
filelock==3.12.4
fsspec==2023.9.2
ftfy==6.1.1
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
lit==17.0.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.22.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pkg_resources==0.0.0
protobuf==3.20.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scipy==1.10.1
sentencepiece==0.1.99
sympy==1.12
timm==0.9.7
tokenizers==0.14.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@21dc5859421cf0d7d82d374b10f533611745a8c5
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
